Some minor renames and extractions
