+++ service-providers/network-requester/src/core.rs
@@ -2,11 +2,11 @@
 // SPDX-License-Identifier: Apache-2.0
 
 use crate::allowed_hosts::{HostsStore, OutboundRequestFilter};
-use crate::connection::Connection;
 use crate::error::NetworkRequesterError;
 use crate::statistics::ServiceStatisticsCollector;
 use crate::websocket;
 use crate::websocket::TSWebsocketStream;
+use crate::{reply, socks5};
 use client_connections::{
     ConnectionCommand, ConnectionCommandReceiver, LaneQueueLengths, TransmissionLane,
 };
@@ -42,56 +42,6 @@ pub struct ServiceProvider {
     stats_provider_addr: Option<Recipient>,
 }
 
-// TODO: move elsewhere after things settle
-#[derive(Debug, Clone)]
-pub enum ReturnAddress {
-    Known(Box<Recipient>),
-    Anonymous(AnonymousSenderTag),
-}
-
-impl From<Recipient> for ReturnAddress {
-    fn from(recipient: Recipient) -> Self {
-        ReturnAddress::Known(Box::new(recipient))
-    }
-}
-
-impl From<AnonymousSenderTag> for ReturnAddress {
-    fn from(sender_tag: AnonymousSenderTag) -> Self {
-        ReturnAddress::Anonymous(sender_tag)
-    }
-}
-
-impl ReturnAddress {
-    fn new(
-        explicit_return_address: Option<Recipient>,
-        implicit_tag: Option<AnonymousSenderTag>,
-    ) -> Option<Self> {
-        // if somehow we received both, always prefer the explicit address since it's way easier to use
-        if let Some(recipient) = explicit_return_address {
-            return Some(ReturnAddress::Known(Box::new(recipient)));
-        }
-        if let Some(sender_tag) = implicit_tag {
-            return Some(ReturnAddress::Anonymous(sender_tag));
-        }
-        None
-    }
-
-    fn send_back_to(self, message: Vec<u8>, connection_id: u64) -> ClientRequest {
-        match self {
-            ReturnAddress::Known(recipient) => ClientRequest::Send {
-                recipient: *recipient,
-                message,
-                connection_id: Some(connection_id),
-            },
-            ReturnAddress::Anonymous(sender_tag) => ClientRequest::Reply {
-                message,
-                sender_tag,
-                connection_id: Some(connection_id),
-            },
-        }
-    }
-}
-
 impl ServiceProvider {
     pub fn new(
         websocket_address: String,
@@ -123,7 +73,7 @@ impl ServiceProvider {
     /// via the `websocket_writer`.
     async fn mixnet_response_listener(
         mut websocket_writer: SplitSink<TSWebsocketStream, Message>,
-        mut mix_reader: MixProxyReader<(Socks5Message, ReturnAddress)>,
+        mut mix_reader: MixProxyReader<(Socks5Message, reply::ReturnAddress)>,
         stats_collector: Option<ServiceStatisticsCollector>,
         mut client_connection_rx: ConnectionCommandReceiver,
     ) {
@@ -245,34 +195,39 @@ impl ServiceProvider {
     async fn start_proxy(
         conn_id: ConnectionId,
         remote_addr: String,
-        return_address: ReturnAddress,
+        return_address: reply::ReturnAddress,
         controller_sender: ControllerSender,
-        mix_input_sender: MixProxySender<(Socks5Message, ReturnAddress)>,
+        mix_input_sender: MixProxySender<(Socks5Message, reply::ReturnAddress)>,
         lane_queue_lengths: LaneQueueLengths,
         shutdown: TaskClient,
     ) {
-        let mut conn =
-            match Connection::new(conn_id, remote_addr.clone(), return_address.clone()).await {
-                Ok(conn) => conn,
-                Err(err) => {
-                    log::error!(
-                        "error while connecting to {:?} ! - {:?}",
-                        remote_addr.clone(),
-                        err
-                    );
+        let mut conn = match socks5::tcp::Connection::new(
+            conn_id,
+            remote_addr.clone(),
+            return_address.clone(),
+        )
+        .await
+        {
+            Ok(conn) => conn,
+            Err(err) => {
+                log::error!(
+                    "error while connecting to {:?} ! - {:?}",
+                    remote_addr.clone(),
+                    err
+                );
 
-                    // inform the remote that the connection is closed before it even was established
-                    mix_input_sender
-                        .send((
-                            Socks5Message::Response(Response::new(conn_id, Vec::new(), true)),
-                            return_address,
-                        ))
-                        .await
-                        .expect("InputMessageReceiver has stopped receiving!");
+                // inform the remote that the connection is closed before it even was established
+                mix_input_sender
+                    .send((
+                        Socks5Message::Response(Response::new(conn_id, Vec::new(), true)),
+                        return_address,
+                    ))
+                    .await
+                    .expect("InputMessageReceiver has stopped receiving!");
 
-                    return;
-                }
-            };
+                return;
+            }
+        };
 
         // Connect implies it's a fresh connection - register it with our controller
         let (mix_sender, mix_receiver) = mpsc::unbounded();
@@ -308,13 +263,14 @@ impl ServiceProvider {
     async fn handle_proxy_connect(
         &mut self,
         controller_sender: &mut ControllerSender,
-        mix_input_sender: &MixProxySender<(Socks5Message, ReturnAddress)>,
+        mix_input_sender: &MixProxySender<(Socks5Message, reply::ReturnAddress)>,
         lane_queue_lengths: LaneQueueLengths,
         sender_tag: Option<AnonymousSenderTag>,
         connect_req: Box<ConnectRequest>,
         shutdown: TaskClient,
     ) {
-        let return_address = match ReturnAddress::new(connect_req.return_address, sender_tag) {
+        let return_address = match reply::ReturnAddress::new(connect_req.return_address, sender_tag)
+        {
             Some(address) => address,
             None => {
                 log::warn!(
@@ -375,7 +331,7 @@ impl ServiceProvider {
         &mut self,
         message: ReconstructedMessage,
         controller_sender: &mut ControllerSender,
-        mix_input_sender: &MixProxySender<(Socks5Message, ReturnAddress)>,
+        mix_input_sender: &MixProxySender<(Socks5Message, reply::ReturnAddress)>,
         lane_queue_lengths: LaneQueueLengths,
         stats_collector: Option<ServiceStatisticsCollector>,
         shutdown: TaskClient,
@@ -441,7 +397,7 @@ impl ServiceProvider {
         // channels responsible for managing messages that are to be sent to the mix network. The receiver is
         // going to be used by `mixnet_response_listener`
         let (mix_input_sender, mix_input_receiver) =
-            tokio::sync::mpsc::channel::<(Socks5Message, ReturnAddress)>(1);
+            tokio::sync::mpsc::channel::<(Socks5Message, reply::ReturnAddress)>(1);
 
         // Used to notify tasks to shutdown. Not all tasks fully supports this (yet).
         let shutdown = task::TaskManager::default();
